2.2-databases-2/orm_migrations/requirements.txt
django~=4.1.6
psycopg2-binary

django-debug-toolbar~=3.8.1